test(stage): refactor SSR and hydration test cases for accurate simulation
- Split SSR and hydration tests into separate files to isolate `jsdom` and `node` environments, preventing inaccurate SSR simulations and React warnings.
- Replace `render` with `hydrateRoot` combined with `renderToString` in hydration tests to correctly validate the actual server-markup hydration path.
